fix(tag): style fix
