+++ bricks/presentational-bricks/package.json
@@ -27,7 +27,7 @@
   },
   "devDependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@next-core/custom-antd-styles": "^1.8.0",
+    "@next-core/custom-antd-styles": "^1.8.1",
     "@next-libs/basic-components": "^2.0.0",
     "@next-libs/clipboard": "^2.0.0",
     "@next-libs/cmdb-utils": "^1.0.0",
